Project structure was changed. Package names updated. Gitignore file updated to ignore targets/ directory
