+++ src/main/java/com/casualTravel/restservice/RestServiceApplication.java
@@ -1,0 +1,13 @@
+package com.casualTravel.restservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RestServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RestServiceApplication.class, args);
+	}
+
+}
