+++ src/main/java/com/casualTravel/restservice/dto/AuthenticationRequest.java
@@ -1,0 +1,15 @@
+package com.casualTravel.restservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationRequest {
+    private String email;
+    private String password;
+}
