+++ src/main/java/com/casualTravel/restservice/dto/AuthenticationResponse.java
@@ -1,0 +1,11 @@
+package com.casualTravel.restservice.dto;
+
+import lombok.*;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationResponse {
+    private String token;
+}
