+++ src/main/java/com/casualTravel/restservice/dto/RegisterRequest.java
@@ -1,0 +1,16 @@
+package com.casualTravel.restservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegisterRequest {
+    private String username;
+    private String email;
+    private String password;
+}
