+++ src/main/java/com/casualTravel/restservice/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.casualTravel.restservice.repository;
+
+import com.casualTravel.restservice.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+    Optional<User> findByEmail (String email);
+}
