--- src/main/java/com/example/restservice/RestServiceApplication.java
@@ -1,13 +1,0 @@
-package com.example.restservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RestServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RestServiceApplication.class, args);
-	}
-
-}
