--- src/main/java/com/example/restservice/dto/AuthenticationResponse.java
@@ -1,11 +1,0 @@
-package com.example.restservice.dto;
-
-import lombok.*;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse {
-    private String token;
-}
