--- src/main/java/com/example/restservice/dto/RegisterRequest.java
@@ -1,16 +1,0 @@
-package com.example.restservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-    private String username;
-    private String email;
-    private String password;
-}
