--- src/main/java/com/example/restservice/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.restservice.repository;
-
-import com.example.restservice.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-    Optional<User> findByEmail (String email);
-}
